compose_example/g1/requirements.txt
fastapi[all]
uvicorn[standard]
gunicorn

strawberry-graphql